requirements.txt
fastapi==0.135.1
jinja2==3.1.6
numpy==2.3.3
openai==2.29.0
opencv-python-headless==4.13.0.92
pillow==12.1.1
pydantic==2.12.0
python-dotenv==1.2.2
python-multipart==0.0.20
torch==2.10.0
transformers==4.57.6
uvicorn[standard]==0.38.0
